requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.49
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
cryptography==42.0.7
sse-starlette==2.1.0
httpx==0.27.0
alembic==1.18.4
pytest==8.2.0
pytest-asyncio==0.23.6
